requirements.txt
IPython
Numpy
pillow
bump2version
pytest
pytest-cov
shortuuid
pandas